requirements.txt
streamlit>=1.28.1
opencv-python>=4.8.1.78
numpy>=1.24.3
scikit-learn>=1.3.2
pandas>=2.1.3
Pillow>=10.1.0
matplotlib>=3.8.2
scikit-image>=0.22.0
scipy>=1.11.4
python-dotenv>=1.0.0
tqdm>=4.66.1
joblib>=1.3.2
seaborn>=0.12.0
torch>=2.0.0
torchvision>=0.15.0
